requirements.txt
appdirs==1.4.4
asgiref==3.6.0
blessed==1.20.0
bpython==0.24
Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
cupshelpers==1.0
curtsies==0.4.1
cwcwidth==0.1.8
dbus-python==1.2.16
distlib==0.3.1
distro==1.5.0
distro-info==1.0
Django==4.2
djangorestframework==3.14.0
filelock==3.0.12
greenlet==2.0.2
gyp==0.1
httplib2==0.18.1
idna==2.10
importlib-metadata==1.6.0
louis==3.16.0
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==4.2.0
pbr==5.5.0
pycairo==1.16.2
pycups==2.0.1
pycurl==7.43.0.6
Pygments==2.14.0
PyGObject==3.38.0
PySimpleSOAP==1.16.2
pysmbc==1.0.23
python-apt==2.2.1
python-debian==0.1.39
python-debianbts==3.1.0
pytz==2023.3
pyxdg==0.27
reportbug==7.10.3+deb11u1
requests==2.25.1
six==1.16.0
sqlparse==0.4.4
stevedore==3.2.2
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.4.0+ds
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
wcwidth==0.2.6
xdg==5
zipp==1.0.0
